Reorganize contents into node/contents.rs
